--- src/resources/resource_users.py
@@ -1,178 +1,0 @@
-import uuid as uuid_
-from abc import abstractmethod
-
-from flask_restful import marshal_with, reqparse
-from rubix_http.exceptions.exception import NotFoundException, BadDataException, UnauthorizedException
-from rubix_http.resource import RubixResource
-from werkzeug.security import check_password_hash
-
-from src.models.device.model_device import DeviceModel
-from src.models.enum import StateType
-from src.models.fcm_server.model_fcm_server import FcmServerModel
-from src.models.user.model_user import UserModel
-from src.resources.rest_schema.schema_user import user_all_attributes, user_return_fields, user_all_fields_with_children
-from src.resources.utils import encrypt_password, parse_user_update, encode_jwt_token, get_access_token, \
-    decode_jwt_token
-
-
-class UsersResourceList(RubixResource):
-    parser = reqparse.RequestParser()
-    for attr in user_all_attributes:
-        parser.add_argument(attr,
-                            type=user_all_attributes[attr]['type'],
-                            required=user_all_attributes[attr].get('required', False),
-                            help=user_all_attributes[attr].get('help', None),
-                            store_missing=False)
-
-    @classmethod
-    @marshal_with(user_all_fields_with_children)
-    def get(cls):
-        return UserModel.find_all()
-
-    @classmethod
-    def post(cls):
-        args = cls.parser.parse_args()
-        uuid = str(uuid_.uuid4())
-        user = UserModel(uuid=uuid, **args)
-        user.password = encrypt_password(user.password)
-        user.save_to_db()
-        return encode_jwt_token(user.uuid, user.username)
-
-
-class UsersResource(RubixResource):
-    @classmethod
-    @marshal_with(user_all_fields_with_children)
-    def get(cls, **kwargs):
-        user: UserModel = cls.get_user(**kwargs)
-        if user is None:
-            raise NotFoundException('User does not exist')
-        return user
-
-    @classmethod
-    @marshal_with(user_return_fields)
-    def patch(cls, **kwargs):
-        args = parse_user_update()
-        user: UserModel = cls.get_user(**kwargs)
-        if user is None:
-            raise NotFoundException("User does not exist")
-        user.update(**args)
-        return user
-
-    @classmethod
-    def delete(cls, **kwargs):
-        user: UserModel = cls.get_user(**kwargs)
-        if user is None:
-            raise NotFoundException("User does not exist")
-        user.delete_from_db()
-        return '', 204
-
-    @classmethod
-    @abstractmethod
-    def get_user(cls, **kwargs) -> UserModel:
-        raise NotImplementedError
-
-
-class UsersResourceByUUID(UsersResource):
-    @classmethod
-    @abstractmethod
-    def get_user(cls, **kwargs) -> UserModel:
-        return UserModel.find_by_uuid(kwargs.get('uuid'))
-
-
-class UsersResourceByUsername(UsersResource):
-    @classmethod
-    @abstractmethod
-    def get_user(cls, **kwargs) -> UserModel:
-        return UserModel.find_by_username(kwargs.get('username'))
-
-
-class UsersVerifyResource(RubixResource):
-    @classmethod
-    def post(cls):
-        parser = reqparse.RequestParser()
-        parser.add_argument('username', type=str, required=True, store_missing=False)
-        args = parser.parse_args()
-        user: UserModel = UserModel.find_by_username(args['username'])
-        if user is None:
-            raise NotFoundException("User does not exist")
-        user.state = StateType.VERIFIED
-        user.commit()
-        DeviceModel.send_notification_by_user_uuid(user.uuid, FcmServerModel.get_key())
-        return {'message': 'User has been verified successfully'}
-
-
-class UsersChangePasswordResource(RubixResource):
-
-    @classmethod
-    def post(cls):
-        access_token = get_access_token()
-        parser = reqparse.RequestParser()
-        parser.add_argument('username', type=str, required=True, store_missing=False)
-        parser.add_argument('password', type=str, required=True, store_missing=False)
-        parser.add_argument('new_password', type=str, required=True, store_missing=False)
-        args = parser.parse_args()
-        username = decode_jwt_token(access_token).get('username', '')
-        user: UserModel = UserModel.find_by_username(username)
-        if user is None:
-            raise NotFoundException("User does not exist")
-        if not (args['username'] == username and check_password_hash(user.password, args['password'])):
-            raise UnauthorizedException("Invalid username or password")
-        user.password = encrypt_password(args['new_password'])
-        user.commit()
-        return {'message': 'Your password has been changed successfully'}
-
-
-class UsersLoginResource(RubixResource):
-    @classmethod
-    def post(cls):
-        parser = reqparse.RequestParser()
-        parser.add_argument('username', type=str, required=True, store_missing=False)
-        parser.add_argument('password', type=str, required=True, store_missing=False)
-        args = parser.parse_args()
-        user = UserModel.find_by_username(args['username'])
-        if user is None:
-            raise NotFoundException('User does not exist')
-        if not check_password_hash(user.password, args['password']):
-            raise BadDataException('username and password combination is incorrect')
-        return encode_jwt_token(user.uuid, user.username)
-
-
-class UsersCheckByUsernameResource(RubixResource):
-    @classmethod
-    def post(cls):
-        parser = reqparse.RequestParser()
-        parser.add_argument('username', type=str, required=True, store_missing=False)
-        args = parser.parse_args()
-        user: UserModel = UserModel.find_by_username(args['username'])
-        return {'exist': bool(user)}
-
-
-class UsersCheckByEmailResource(RubixResource):
-    @classmethod
-    def post(cls):
-        parser = reqparse.RequestParser()
-        parser.add_argument('email', type=str, required=True, store_missing=False)
-        args = parser.parse_args()
-        user: UserModel = UserModel.find_by_email(args['email'])
-        return {'exist': bool(user)}
-
-
-class UsersRefreshToken(RubixResource):
-    @classmethod
-    def get(cls):
-        access_token = get_access_token()
-        username = decode_jwt_token(access_token).get('username', '')
-        uuid = decode_jwt_token(access_token).get('sub', '')
-        return encode_jwt_token(uuid, username)
-
-
-class UsersAuthenticateResource(RubixResource):
-    @classmethod
-    def get(cls):
-        access_token = get_access_token()
-        username = decode_jwt_token(access_token).get('username', '')
-        uuid = decode_jwt_token(access_token).get('sub', '')
-        user: UserModel = UserModel.find_by_uuid(uuid)
-        if user is None:
-            raise NotFoundException("User does not exist")
-        return {'username': username, 'state': user.state.value}
