+++ src/rest_schema/schema_device.py
